Completed value Added Features and re-factors , Also done code with rearrange

+++ HotelListing.API/Controllers/CountriesV2Controller.cs
@@ -11,43 +11,35 @@ using AutoMapper;
 using HotelListing.API.Core.Contracts;
 using Microsoft.AspNetCore.Authorization;
 using HotelListing.API.Core.Exceptions;
-using HotelListing.API.Core.Models;
+using Microsoft.AspNetCore.OData.Query;
 
 namespace HotelListing.API.Core.Controllers
 {
     [Route("api/v{version:apiVersion}/countries")]
     [ApiController]
-    [ApiVersion("1.0", Deprecated = true)]
-    public class CountriesController : ControllerBase
+    [ApiVersion("2.0")]
+    public class CountriesV2Controller : ControllerBase
     {
         private readonly IMapper _mapper;
         private readonly ICountriesRepository _countriesRepository;
-        private readonly ILogger<CountriesController> _logger;
+        private readonly ILogger<CountriesV2Controller> _logger;
 
-        public CountriesController(IMapper mapper, ICountriesRepository countriesRepository, ILogger<CountriesController> logger)
+        public CountriesV2Controller(IMapper mapper, ICountriesRepository countriesRepository, ILogger<CountriesV2Controller> logger)
         {
             this._mapper = mapper;
             this._countriesRepository = countriesRepository;
             this._logger = logger;
         }
 
-        // GET: api/Countries/GetAll
-        [HttpGet("GetAll")]
-        public async Task<ActionResult<PagedResult<GetCountryDto>>> GetCountries()
-        {
-            //Select * from Countries
-            var countries = await _countriesRepository.GetAllAsync<GetCountryDto>();
-            return Ok(countries);
-        }
-
-        // GET: api/Countries/?StartIndex=0&pagesize=25&PageNumber=1
+        // GET: api/Countries
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<GetCountryDto>>> GetPagedCountries([FromQuery] QueryParameters queryParameters)
-
+        [EnableQuery]
+        public async Task<ActionResult<IEnumerable<GetCountryDto>>> GetCountries()
         {
             //Select * from Countries
-            var pageCountriesResult = await _countriesRepository.GetAllAsync<GetCountryDto>(queryParameters);
-            return Ok(pageCountriesResult);
+            var countries = await _countriesRepository.GetAllAsync();
+            var records = _mapper.Map<List<GetCountryDto>>(countries);
+            return Ok(records);
         }
 
         // GET: api/Countries/5
@@ -55,7 +47,15 @@ namespace HotelListing.API.Core.Controllers
         public async Task<ActionResult<CountryDto>> GetCountry(int id)
         {
             var country = await _countriesRepository.GetDetails(id);
-            return Ok(country);
+
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountries), id);
+            }
+
+            var countryDto = _mapper.Map<CountryDto>(country);
+
+            return Ok(countryDto);
         }
 
         // PUT: api/Countries/5
@@ -69,9 +69,20 @@ namespace HotelListing.API.Core.Controllers
                 return BadRequest("Invalid Record Id");
             }
 
+            //_context.Entry(country).State = EntityState.Modified;
+
+            var country = await _countriesRepository.GetAsync(id);
+
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountries), id);
+            }
+
+            _mapper.Map(updateCountryDto, country);
+
             try
             {
-                await _countriesRepository.UpdateAsynce(id, updateCountryDto);
+                await _countriesRepository.UpdateAsynce(country);
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -92,10 +103,21 @@ namespace HotelListing.API.Core.Controllers
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
         [Authorize]
-        public async Task<ActionResult<CountryDto>> PostCountry(CreateCountryDto createCountryDto)
+        public async Task<ActionResult<Country>> PostCountry(CreateCountryDto createCountryDto)
         {
-            var country = await _countriesRepository.AddAsync<CreateCountryDto, GetCountryDto>(createCountryDto);
-            return CreatedAtAction(nameof(GetCountry), new { id = country.Id }, country);
+            /*var countryOld = new Country
+            {
+                Name = createCountryDto.Name,
+                ShortName = createCountryDto.ShortName
+            };
+            */
+
+            //Using Mapper
+            var country = _mapper.Map<Country>(createCountryDto);
+
+            await _countriesRepository.AddAsync(country);
+
+            return CreatedAtAction("GetCountry", new { id = country.Id }, country);
         }
 
         // DELETE: api/Countries/5
@@ -103,6 +125,12 @@ namespace HotelListing.API.Core.Controllers
         [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> DeleteCountry(int id)
         {
+            var country = await _countriesRepository.GetAsync(id);
+            if (country == null)
+            {
+                throw new NotFoundException(nameof(GetCountries), id);
+            }
+
             await _countriesRepository.DeleteAsync(id);
 
             return NoContent();
